增加ES存储支持

+++ Model/BlogInfo.cs
@@ -9,8 +9,8 @@ namespace Model
     [ElasticsearchType(IdProperty ="Id",Name ="bloginfo")]
     public class BlogInfo
     {
-        [Number(NumberType.Integer)]
-        public int Id { get; set; }
+        [Number(NumberType.Long)]
+        public long Id { get; set; }
         public string Title { get; set; }
         public string Content { get; set; }
         public DateTime CreatedTime { get; set; }
